test(apply): isolate live-state regression
